requirements.txt
discord.py
psycopg
dateparser
python-dotenv
reactionmenu